`xspectra`: Add `get_core_hole_inputs` Utility Function
Adds a new utility function `get_core_hole_inputs` which returns a
`parameters` dict object with suggested inputs for the core-hole
treatment option selected. This is intended to replace the existing
"protocols"-type system with one which suggests suitable inputs based on
the given `parameters` dictionary and information about the existing
magnetic structure (if present). This will cover cases for spin-neutral,
ferromagnetic (FM), and anti-ferromagnetic (AFM) systems.

Note that the new system does not include "none" as an option, since it
would make more sense to skip the function call entirely if a
"no-core-hole" model is requested and also serves as a way for more
advanced users to have full control over parameter inputs.